pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alma-tv"
version = "0.1.0"
description = "Raspberry Pi automation suite for scheduling and playing curated children's programming"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Alma TV Contributors"}
]
keywords = ["raspberry-pi", "media", "automation", "children", "scheduling"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    "sqlalchemy>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "watchdog>=3.0.0",
    "python-vlc>=3.0.0",
    "jinja2>=3.1.0",
    "svgwrite>=1.4.0",
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.12.0",
    "python-fasthtml>=0.12.35",
]

[project.optional-dependencies]
dev = [
    "nbdev>=2.3.12",
    "jupyter>=1.0.0",
    "notebook>=6.0.0",
    "pytest>=7.4.0",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "playwright>=1.40.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
]

[project.scripts]
alma = "alma_tv.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "benchmark: Benchmark tests",
    "hardware: Hardware-dependent tests (skip in CI)",
]

[tool.coverage.run]
source = ["src/alma_tv"]
omit = ["*/tests/*", "*/benchmarks/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
